+++ js/const.js
@@ -1,0 +1,45 @@
+const NAMES = [
+  'Кузьма',
+  'Тихон',
+  'Вилка',
+  'Шурик',
+  'Милана',
+  'Юрец',
+  'Боня',
+  'Мяус',
+  'Просто Кот',
+  'Мирон',
+  'Сева',
+  'Тимон',
+  'Пумба'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Хештег за любовь',
+  'Этот прекрасный мир',
+  'Просто Мяу!',
+  'ЗаМУРчательно',
+  'Жду ваший комментов',
+  'Без слов',
+  'Одни эмоции',
+];
+
+const MAXAVATARNUMBER = 6;
+const MAXPHOTOID = 25;
+const MAXPHOTOURL = 25;
+const MINLIKES = 15;
+const MAXLIKES = 200;
+const MAXCOMMENTS = 30;
+
+const keksId = [25];
+
+export { NAMES, MESSAGES, DESCRIPTIONS, MAXAVATARNUMBER, MAXPHOTOID, MAXPHOTOURL, MINLIKES, MAXLIKES, MAXCOMMENTS, keksId };
