+++ js/util.js
@@ -1,0 +1,22 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomIdentifier = (a = 0,b = 1000)=>{
+  const idNumber = [];
+  return function (){
+    let randomId = 0;
+    do {
+      randomId = getRandomInteger(a,b);
+    } while (idNumber.includes(randomId));
+    idNumber.push(randomId);
+    return randomId;
+  };
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export { getRandomInteger, getRandomIdentifier, getRandomArrayElement, };
